backend/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# API Clients
requests>=2.31.0
httpx>=0.25.0

# Audio Processing
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.3

# Utilities
pydantic>=2.4.2
starlette>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4

# CORS
fastapi-cors>=0.0.6